- Set project compatibility to XCode 3.2. - Enabled x86_64 build architecture. Default is now a 64 bit build of Pecunia. Adjusted a few code places for this switch. - Set LLVM GCC 4.2 as compiler for Pecunia (which is the system default). - Removed SM2D graphics framework. No longer used. - Removed MCEMPieChartView too, as it is no longer used as well. - Removed AqBanking.zip, not used currently. - Updated coreplot to the latest version. - Recompiled HBCIServer.jar and updated the copy in the client folder. - Switched category reporting from SM2D to coreplot. This view is now embedded as subview of the accounts + category section. - BankingController: some code cleanup. - Removed deprecated methods from ShortDate and changed all places where they were used. - GraphicAdditions: better random color computation (and a bug fix), added CGColor method to NSColor (via category). - ImageAndTextCell: code cleanup, disabled state

+++ client/Source/BankStatementController.m
@@ -147,7 +147,7 @@
 	}
 	
 	// assign to category
-	int idx = [categoriesController selectionIndex ];
+	NSInteger idx = [categoriesController selectionIndex ];
 	if (idx != NSNotFound) {
 		Category *cat = [[categoriesController arrangedObjects ] objectAtIndex:idx];
 		[newStatement assignToCategory:cat ];
